requirements.txt
boto3==1.42.57
python-dotenv==1.2.1

pydantic-settings~=2.13.1
pydantic~=2.12.5
botocore~=1.42.57
mcp~=1.23.0